created component framework
