+++ .env.default
@@ -4,4 +4,4 @@ DB_USER=
 DB_PASSWORD=
 DB_NAME=
 
-TOKEN="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"
+TOKEN="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"
