+++ .gitignore
@@ -1,3 +1,3 @@
 node_modules/
-.env
+.env.default
 package-lock.json
