+++ client/src/components/relatedItems/ComparisonModal.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+//import proptype later
+
+class ComparisonModal extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {}
+  }
+}
+
+export default ComparisonModal;
