+++ client/src/components/relatedItems/RelatedItemsCard.jsx
@@ -1,0 +1,12 @@
+import React from 'react.jsx';
+//import proptype later
+
+class RelatedItemsCard extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {}
+  }
+
+}
+
+export default RelatedItemsCard;
