+++ client/src/components/relatedItems/RelatedItemsCarousel.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import RelatedItemsCard from './RelatedItemsCard.jsx';
+//import proptype later
+
+class RelatedItemsCarousel extends React.Component {
+  constructor(props) {
+    super(props);
+    this.state = {}
+  }
+}
+
+export default RelatedItemsCarousel;
