+++ server/helper.js
@@ -5,7 +5,7 @@ exports.configureOptions = (req) => {
     },
     url: "https://app-hrsei-api.herokuapp.com/api/fec2/rfp" + req.url,
     method: req.method,
-    params: req.query,
+    // params: req.query,
     data: req.body
   }
   // console.log(req.method)
@@ -14,3 +14,5 @@ exports.configureOptions = (req) => {
   // console.log(options, '<<<<<<<<<<<<<<<<<<<<<<')
   return options;
 };
+
+//req.url adds all the params at the end...redundant
